requirements.txt
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.30.0
huggingface-hub>=0.23.0
sentence-transformers>=2.7.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
langchain-chroma>=0.2.0
pypdf>=4.2.0
pdfplumber>=0.11.0
langchain-huggingface
python-dotenv>=1.0.1
tqdm>=4.66.0
streamlit>=1.37.0
gdown>=5.2.0
google-genai>=0.1.0
pandas>=2.2.0
scikit-learn>=1.5.0
python-docx>=1.1.2
openpyxl>=3.1.2
